% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_env.R
\name{check_course_tree}
\alias{check_course_tree}
\title{Validate a course/week bundle is installed in the package}
\usage{
check_course_tree(course = "mc451", week = 1)
}
\description{
Validate a course/week bundle is installed in the package
}
\keyword{internal}
